pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "music-catalog-pe-env"
version = "0.1.0"
description = "OpenEnv environment for AI-driven private equity analysis of music catalogs"
authors = [
    { name = "ishikat" }
]
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "openai>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.setuptools]
include-package-data = true
packages = ["music_catalog_pe_env", "server"]

[project.scripts]
server = "server.app:main"
